# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kampala Bakery Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
...
  /inventory/{productId}/adjust:
    post:
      operationId: adjustInventory
      tags: [inventory]
      summary: Manually adjust inventory
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustInventoryBody"
      responses:
        "200":
          description: Adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"

  # SALES (POS)
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new customer order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Add employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ATTENDANCE
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance/check-in:
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Employee check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInBody"
      responses:
        "201":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/{id}/check-out:
    put:
      operationId: checkOut
      tags: [attendance]
      summary: Employee check-out
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # PAYMENTS
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, staff, rider]
        employeeId:
          type: integer
          nullable: true
      required: [id, username, name, role]
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        currentStock:
          type: integer
        lowStockThreshold:
          type: integer
        isLow:
          type: boolean
        lastUpdated:
          type: string
      required: [id, productId, productName, currentStock, lowStockThreshold, isLow, lastUpdated]
    AdjustInventoryBody:
      type: object
      properties:
        quantity:
          type: integer
        reason:
          type: string
      required: [quantity, reason]
...
    SaleItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
      required: [productId, productName, quantity, unitPrice, subtotal]
...
    SaleItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryLocation:
          type: string
        status:
          type: string
          enum: [pending, confirmed, out_for_delivery, delivered, cancelled]
        totalAmount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, mtn_momo, airtel_money]
        transactionId:
          type: string
          nullable: true
        placedAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
      required: [id, customerName, customerPhone, deliveryLocation, status, totalAmount, paymentMethod, placedAt, items]
    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryLocation:
          type: string
        paymentMethod:
          type: string
          enum: [cash, mtn_momo, airtel_money]
        transactionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
      required: [customerName, customerPhone, deliveryLocation, paymentMethod, items]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, out_for_delivery, delivered, cancelled]
      required: [status]
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
          enum: [admin, baker, cashier, rider]
        phone:
          type: string
        email:
          type: string
          nullable: true
        salary:
          type: number
          nullable: true
        joinDate:
          type: string
        isActive:
          type: boolean
      required: [id, name, role, phone, joinDate, isActive]
    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, baker, cashier, rider]
        phone:
          type: string
        email:
          type: string
        salary:
          type: number
        joinDate:
          type: string
        isActive:
          type: boolean
      required: [name, role, phone]
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
          nullable: true
        date:
          type: string
        hoursWorked:
          type: number
          nullable: true
      required: [id, employeeId, employeeName, checkIn, date]
    CheckInBody:
      type: object
      properties:
        employeeId:
          type: integer
      required: [employeeId]
...